Remove website from this repo and rework to remove /store and /edit routes
