--- app/page.module.css
@@ -1,64 +1,0 @@
-.container {
-	display: flex;
-	justify-content: center;
-}
-
-.page {
-	max-width: 60em;
-	padding-bottom: 6.5em;
-}
-
-.splash {
-	height: 100vh;
-	display: flex;
-	align-items: center;
-}
-
-.name {
-	color: var(--color-accent-100);
-}
-
-.page h1 {
-	font-size: 4.5rem;
-	font-weight: 900;
-}
-
-.page h2 {
-	margin: 1.35em 0 0 0;
-	font-size: 2.5rem;
-	font-weight: 900;
-}
-
-.info {
-	margin-bottom: 3.5em;
-	font-size: 2.35rem !important;
-	line-height: 2.85rem !important;
-	font-style: italic;
-}
-
-.info:last-of-type {
-	margin: 1.5em 0;
-}
-
-.page p {
-	font-size: 1.35rem;
-	line-height: 1.85rem;
-}
-
-.image {
-	width: 100%;
-	height: auto;
-	margin: 1.15em 0 2.5em 0;
-}
-
-.page ol,
-.page ul {
-	margin-left: 0.5em;
-	padding-left: 0.65rem;
-	font-size: 1.35rem;
-	line-height: 2.35rem;
-}
-
-.page li {
-	margin: 0.35em 0;
-}
